Dev branch

+++ MiniERP.PurchaseOrderService/Models/PODetail.cs
@@ -6,8 +6,8 @@
         public int OrderNo { get; set; }
         public int LineNo { get; set; }
         public int Productd { get; set; }
-        public decimal Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public double Quantity { get; set; }
+        public double UnitPrice { get; set; }
         public string Notes { get; set; } = string.Empty;
     }
 }
